% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastScatter.R
\name{blastScatter}
\alias{blastScatter}
\title{Scatter plot of slcp-formatted blastn results}
\usage{
blastScatter(x, showbreaks = F, strandcolor = F, border = NA, ...)
}
\description{
Can work on multiple query contigs and subject hits.
}
