% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastScatterWindow.R
\name{blastScatterWindow}
\alias{blastScatterWindow}
\title{Blastn scatter plot with windowing}
\usage{
blastScatterWindow(x, s = 1, e = 10000, border = NA, min.id = 0.9,
  ...)
}
\description{
Plot slcp-formatted blastn data frame, can do windowing.
}
