% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contigDetail.R
\name{contigDetail}
\alias{contigDetail}
\title{Return details of blast results for a single contig (query)}
\usage{
contigDetail(x, meta = ref_meta)
}
\description{
Requires a data frame of slcp-formatted blastn results for only one contig
from one query genome. This might be useful for example for known complete
plasmid or chromosome on a single contig.
}
