% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contigSummary.R
\name{contigSummary}
\alias{contigSummary}
\title{Summarize blast hits for a given contig (query)}
\usage{
contigSummary(x, meta = ref_meta)
}
\description{
Needs slcp-formatted data frame with data from only one contig from one
query genome.
}
\details{
Internal procedure, generally used in conjunction with ddply.
}
