% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDetail.R
\name{plotDetail}
\alias{plotDetail}
\title{Plot to visualize `contigDetail` or `subjectDetail`}
\usage{
plotDetail(x, classification = NULL, border = NULL, minid = 0.7, ...)
}
\description{
Plot to visualize `contigDetail` or `subjectDetail`
}
