% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackPlot.R
\name{stackPlot}
\alias{stackPlot}
\title{Stacked bar graph of contig classifications}
\usage{
stackPlot(x, border = NA, col = c("gray", "red", "blue"), ...)
}
\description{
Stacked bar graph of contig classifications
}
