% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subjectDetail.R
\name{subject_detail}
\alias{subject_detail}
\title{Return details of blast results for a single subject (database/reference)}
\usage{
subject_detail(x)
}
\description{
Requires a data frame of slcp-formatted blastn results for only one subject
that was hit. This might be useful for example for checking coverage of a
known reference plasmid from Genbank
}
